// app/backend/src/modules/analytics/constants/event-types.ts
export enum AnalyticsEventType {
  PAGE_VIEW = 'page_view',
  VIEW_CATEGORY = 'view_category',
  VIEW_PRODUCT = 'view_product',
  VIEW_POST = 'view_post',
  ADD_TO_CART = 'add_to_cart',
  REMOVE_FROM_CART = 'remove_from_cart',
  CHECKOUT_START = 'checkout_start',
  PAYMENT_SUCCESS = 'payment_success',
  DWELL_TIME = 'dwell_time',
}